requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
certifi==2024.8.30
click==8.1.7
deprecation==2.1.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.0
fastapi-cli==0.0.5
frozenlist==1.4.1
gotrue==2.9.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.0
mdurl==0.1.2
multidict==6.1.0
packaging==24.1
postgrest==0.17.1
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.2
realtime==2.0.5
rich==13.9.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
storage3==0.8.1
StrEnum==0.4.15
supabase==2.9.0
supafunc==0.6.1
typer==0.12.5
typing_extensions==4.12.2
uvicorn==0.31.0
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1
yarl==1.14.0
